+++ test/ambiance_test.dart
@@ -8,7 +8,8 @@ void main() {
   group('Ambiance tests', () {
     test('Initialisation', () {
       final a = Ambiance(
-          SoundReference('sound.wav', SoundType.file), Point(5.0, 4.0));
+          sound: SoundReference('sound.wav', SoundType.file),
+          position: Point(5.0, 4.0));
       expect(a.sound.name, equals('sound.wav'));
       expect(a.sound.type, equals(SoundType.file));
       expect(a.position, equals(Point<double>(5.0, 4.0)));